Split out external address (#86)
This splits out the external address from the wallet address.

This exposes the `balances`, `balance`, and `faucet` functionality on
the external address.

```
address = Coinbase::ExternalAddress.new(:base_sepolia, '0x12345')
address.balances
```

<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
