+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Support external addresses for balance fetching and requesting faucet funds.
+
 ## [0.0.8] - 2024-06-18
 
 ### Added
